Move getDateFromCreditCardExpiry to utils/datetime
