+++ client/dashboard/me/purchases/util.ts
@@ -1,4 +1,4 @@
-import { isWithinLast, isWithinNext } from '../../utils/datetime';
+import { isWithinLast, isWithinNext, getDateFromCreditCardExpiry } from '../../utils/datetime';
 import { SubscriptionBillPeriod, AkismetPlans } from './constants';
 import type { Purchase } from '../../data/purchase';
 
@@ -73,34 +73,6 @@ export function isCloseToExpiration( purchase: Purchase ): boolean {
 			? SubscriptionBillPeriod.PLAN_MONTHLY_PERIOD
 			: SubscriptionBillPeriod.PLAN_MONTHLY_PERIOD * 3;
 	return isWithinNext( new Date( purchase.expiry_date ), threshold, 'days' );
-}
-
-/**
- * Transforms a credit card expiry date like `11/23` into a Date representing 2023-11-30.
- *
- * Returns the last day of the month because credit cards typically expire at the end of the valid month.
- */
-function getDateFromCreditCardExpiry( cardExpiryDate: string ): Date {
-	const [ month, year ] = cardExpiryDate.split( '/' );
-	if ( ! month || ! year ) {
-		throw new Error( `Could not parse credit card date '${ cardExpiryDate }'` );
-	}
-	const monthNumber = parseInt( month );
-	const yearNumber = parseInt( year );
-	if ( isNaN( monthNumber ) || isNaN( yearNumber ) ) {
-		throw new Error( `Could not parse credit card date '${ cardExpiryDate }'` );
-	}
-	const currentYear = new Date().getFullYear();
-	const currentMillenniumPrefix = Math.floor( currentYear / 100 );
-	const fullYear = parseInt( `${ currentMillenniumPrefix }${ yearNumber }` );
-	if ( isNaN( fullYear ) ) {
-		throw new Error( `Could not parse credit card date '${ cardExpiryDate }'` );
-	}
-	// Note that the Date constructor expects the month to be 0 indexed, so 0
-	// is January, but specifying a 0 as the day "underflows" and goes to the
-	// last day of the previous month, which allows us to pass the wrong index
-	// and get the right result.
-	return new Date( fullYear, monthNumber, 0 );
 }
 
 export function creditCardExpiresBeforeSubscription( purchase: Purchase ): boolean {
